+++ deployToLinode.sh
@@ -51,41 +51,19 @@ log $HR
 log "Deploying ${APPNAME} to ${USERHOST}:/${TARGETDIR}"
 log $HR
 log ""
+log "Adding remote installation script into deploy directory"
 
-if [ "$DLLNAME" != "" ]; then
-    log "Writing supervisor configuration file into deploy directory"
-
-    cat > ${SOURCEDIR}/${APPNAME}.conf <<EOF
-    [program:${APPNAME}]
-    command=/usr/bin/dotnet ${TARGETDIR}/${DLLNAME}.dll
-    directory=${TARGETDIR}
-    autostart=true
-    autorestart=true
-    stderr_logfile=/var/log/${APPNAME}.err.log
-    stdout_logfile=/var/log/${APPNAME}.out.log
-    environment=ASPNETCORE_ENVIRONMENT=Production
-    user=deploy
-    stopsignal=INT
-EOF
-
-    log "Adding remote installation script into deploy directory"
-
-    cat > ${SOURCEDIR}/localInstall.sh <<EOF2
-    #!/bin/bash
-    cp ${TARGETDIR}/${APPNAME}.conf /etc/supervisor/conf.d
-    sudo supervisorctl reread
-    sudo supervisorctl update
+cat > ${SOURCEDIR}/localInstall.sh <<EOF2
+#!/bin/bash
+ln -s ${TARGETDIR}/../repl ${TARGETDIR}/repl
 EOF2
-fi
 
 log "Sending deploy directory to remote instance"
 
 cout rsync -avz -e 'ssh' ${SOURCEDIR}/. ${USERHOST}:${TARGETDIR} --delete
 
-if [ "$DLLNAME" != "" ]; then
-    log "Installing application under supervisor and starting"
-    cout ssh ${USERHOST} /bin/bash ${TARGETDIR}/localInstall.sh
-fi
+log "Executing localInstall.sh"
+cout ssh ${USERHOST} /bin/bash ${TARGETDIR}/localInstall.sh
 
 log ""
 log "Deployment to ${USERHOST} completed"