+++ src/App/TransitionCustom.fs
@@ -34,10 +34,6 @@ let view() =
     let visible = Store.make false
 
     Html.div [
-        disposeOnUnmount [visible]
-
-        onMount (fun _ -> visible <~ true) []
-
         class' "container"
 
         Html.label [
@@ -52,4 +48,7 @@ let view() =
             Html.p [
                 text "The quick brown fox jumps over the lazy dog"
             ]
+
+        disposeOnUnmount [visible]
+        onMount (fun _ -> true |> Store.set visible) [] // Force a transition upon first showing
     ]